move chat to root page
